Chore: 함수 분리:
