+++ src/pages/home/index.tsx
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { themedPalette } from "@/styles/theme";
 import { PageContainer } from "@/styles/common.style";
 
 const Home = () => {
@@ -31,21 +30,7 @@ const Wrapper = styled.div`
   line-height: 1.9rem;
 `;
 
-const KakaoButton = styled.button`
-  width: 180px;
-  height: 50px;
-  margin-top: 50px;
-  display: flex;
-  align-items: center;
-  border-radius: 10px;
-  background: url("http://ec2-3-39-24-149.ap-northeast-2.compute.amazonaws.com:8080/images/kakao_login.png");
-  background-repeat: no-repeat;
-  object-fit: cover;
-  cursor: pointer;
-`;
-
 const S = {
   Container,
   Wrapper,
-  KakaoButton,
 };
