+++ src/pages/index.tsx
@@ -3,34 +3,16 @@ import { PageContainer } from "@/styles/common.style";
 import { themedPalette } from "@/styles/theme";
 import Image from "next/image";
 import kakao_logo from "../../public/kakao.png";
-import { useRouter } from "next/router";
 import Head from "next/head";
+import { movePage } from "@/util/movePage";
 
 const Home = () => {
-  const router = useRouter();
   const KAKAO_REST_API_KEY = process.env.NEXT_PUBLIC_KAKAO_CLIENT_ID;
   const KAKAO_REDIRECT_URI = process.env.NEXT_PUBLIC_KAKAO_REDIRECT_URI;
   const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${KAKAO_REST_API_KEY}&redirect_uri=${KAKAO_REDIRECT_URI}&response_type=code`;
 
-  const kakaoLogin = async () => {
+  const kakaoLogin = () => {
     window.location.href = KAKAO_AUTH_URL;
-  };
-
-  const movePage = (type: "member" | "study" | "study_rank" | "member_rank") => {
-    switch (type) {
-      case "member":
-        router.push("/member/1");
-        return;
-      case "study":
-        router.push("/study/1");
-        return;
-      case "study_rank":
-        router.push("/rank/study");
-        return;
-      case "member_rank":
-        router.push("/rank/algorithm");
-        return;
-    }
   };
 
   return (
