+++ src/pages/study/[studyId]/mission/[missionId].tsx
@@ -29,8 +29,6 @@ const MissionDetail = () => {
   const missionId = router.query.missionId;
   const { data: missionData, isLoading: getMissionDataLoading } = studyApi.useGetStudyRuleQuery(missionId);
   const { data: memberList, isLoading: getMemberListLoading } = studyApi.useGetStudyMemberListQuery(studyId);
-  console.log(missionData);
-  console.log(memberList);
 
   const [memberStatus, setMemberStatus] = useState<IPersonStudyRuleDtst[]>();
   const [missionStatus, setMissionState] = useState<string>();
